+++ bindings/core/tests/secrets_debug.rs
@@ -1,9 +1,10 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+use iota_sdk::client::secret::SecretManagerDto;
 #[cfg(feature = "ledger_nano")]
 use iota_sdk::client::secret::SecretManagerDto;
-use iota_sdk_bindings_core::{ClientMethod, Response, WalletMethod};
+use iota_sdk_bindings_core::{ClientMethod, Response, WalletMethod, WalletOptions};
 
 #[test]
 fn method_interface_secrets_debug() {
@@ -43,4 +44,15 @@ fn method_interface_secrets_debug() {
 
     let response = Response::GeneratedMnemonic("mnemonic".to_string());
     assert_eq!(format!("{:?}", response), "GeneratedMnemonic(<omitted>)");
+
+    let wallet_options = WalletOptions {
+        storage_path: None,
+        client_options: None,
+        coin_type: None,
+        secret_manager: Some(SecretManagerDto::Placeholder),
+    };
+    assert_eq!(
+        format!("{:?}", wallet_options),
+        "WalletOptions { storage_path: None, client_options: None, coin_type: None, secret_manager: Some(<omitted>) }"
+    );
 }
